edits database tables and adds #has_secure_password to Customer and Truck models

+++ db/migrate/20210328211131_create_customers.rb
@@ -4,7 +4,7 @@ class CreateCustomers < ActiveRecord::Migration[6.1]
       t.string :name
       t.string :password_digest
       t.integer :zipcode
-      t.integer :wallet
+      t.float :wallet
     end
   end
 end
